+++ src/main/java/org/jglr/sbm/types/IntType.java
@@ -1,10 +1,10 @@
 package org.jglr.sbm.types;
 
 public class IntType extends Type {
-    private final int width;
+    private final long width;
     private final boolean isSigned;
 
-    public IntType(int width, boolean isSigned) {
+    public IntType(long width, boolean isSigned) {
         super((isSigned ? "" : "u")+"int"+width);
         this.width = width;
         this.isSigned = isSigned;
@@ -14,7 +14,7 @@ public class IntType extends Type {
         return isSigned;
     }
 
-    public int getWidth() {
+    public long getWidth() {
         return width;
     }
 }
