+++ src/main/java/org/jglr/sbm/types/VectorType.java
@@ -2,15 +2,15 @@ package org.jglr.sbm.types;
 
 public class VectorType extends Type {
     private final Type componentType;
-    private final int componentCount;
+    private final long componentCount;
 
-    public VectorType(Type componentType, int componentCount) {
+    public VectorType(Type componentType, long componentCount) {
         super("vec"+componentCount+"("+componentType+")");
         this.componentType = componentType;
         this.componentCount = componentCount;
     }
 
-    public int getComponentCount() {
+    public long getComponentCount() {
         return componentCount;
     }
